+++ packages/component-library/src/I18nProvider/index.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import type { ComponentProps } from "react";
+import { I18nProvider as I18nProviderBase } from "react-aria";
+
+export const I18nProvider = (
+  props: Omit<ComponentProps<typeof I18nProviderBase>, "locale">,
+) => {
+  return <I18nProviderBase locale="en-US" {...props} />;
+};