+++ services/paymaster/src/constraint/templates.rs
@@ -159,8 +159,8 @@ impl InstructionConstraint {
                 // instruction = 1 (CreateIdempotent)
                 DataConstraint {
                     start_byte: 0,
-                    data_type: DataType::U32,
-                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U32(1)]),
+                    data_type: DataType::U8,
+                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U8(1)]),
                 },
             ],
             required: false,
@@ -176,8 +176,8 @@ impl InstructionConstraint {
                 // instruction = 17 (SyncNative)
                 DataConstraint {
                     start_byte: 0,
-                    data_type: DataType::U32,
-                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U32(17)]),
+                    data_type: DataType::U8,
+                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U8(17)]),
                 },
             ],
             required: false,
@@ -193,8 +193,8 @@ impl InstructionConstraint {
                 // instruction = 9 (CloseAccount)
                 DataConstraint {
                     start_byte: 0,
-                    data_type: DataType::U32,
-                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U32(9)]),
+                    data_type: DataType::U8,
+                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U8(9)]),
                 },
             ],
             required: false,