+++ packages/sessions-sdk-ts/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @fogo/sessions-sdk
 
+## 0.0.32 (2026-01-08)
+
+### Fixes
+
+- Use v4 sessions
+
+## 0.0.30 (2026-01-08)
+
+### Fixes
+
+- Add getSessionWrapInstructions and getSessionUnwrapInstructions to allow wrapping and unwrapping native tokens within sessions
+- Support transferring native tokens with the intent transfer program
+
+## 0.0.29 (2026-01-05)
+
+### Fixes
+
+- Fix ledger support for new firmware (for wallets that have support for signing messages on new Ledger firmware)
+
 ## 0.0.28 (2025-12-10)
 
 ### Fixes